Stop using base class in test to avoid reusing endpoint names
