+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -321,7 +321,6 @@
     <Compile Include="Recoverability\When_immediate_retries_disabled.cs" />
     <Compile Include="Recoverability\When_transactions_off.cs" />
     <Compile Include="Sagas\when_reply_from_saga_not_found_handler.cs" />
-    <Compile Include="Sagas\When_doing_request_response_between_sagas.cs" />
     <Compile Include="Sagas\when_receiving_multiple_timeouts.cs" />
     <Compile Include="Sagas\When_saga_is_mapped_to_complex_expression.cs" />
     <Compile Include="Sagas\When_saga_has_a_non_empty_constructor.cs" />
