+++ src/NServiceBus.AcceptanceTests/Sagas/When_doing_request_response_between_sagas_response_from_noninitiating.cs
@@ -1,20 +1,132 @@
 namespace NServiceBus.AcceptanceTests.Sagas
 {
+    using System;
     using System.Threading.Tasks;
     using AcceptanceTesting;
+    using EndpointTemplates;
+    using Features;
     using NUnit.Framework;
 
-    public class When_doing_request_response_between_sagas_response_from_noninitiating : When_doing_request_response_between_sagas
+    public class When_doing_request_response_between_sagas_response_from_noninitiating : NServiceBusAcceptanceTest
     {
         [Test]
-        public async Task Should_autocorrelate_the_response_back_to_the_requesting_saga_from_handler_other_than_the_initiating_one()
+        public async Task Should_autocorrelate_the_response_back()
         {
-            var context = await Scenario.Define<Context>(c => { c.ReplyFromNonInitiatingHandler = true; })
+            var context = await Scenario.Define<Context>()
                 .WithEndpoint<Endpoint>(b => b.When(session => session.SendLocal(new InitiateRequestingSaga())))
                 .Done(c => c.DidRequestingSagaGetTheResponse)
                 .Run();
 
             Assert.True(context.DidRequestingSagaGetTheResponse);
+        }
+
+        public class Context : ScenarioContext
+        {
+            public bool DidRequestingSagaGetTheResponse { get; set; }
+        }
+
+        public class Endpoint : EndpointConfigurationBuilder
+        {
+            public Endpoint()
+            {
+                EndpointSetup<DefaultServer>(config => config.EnableFeature<TimeoutManager>());
+            }
+
+            public class RequestResponseRequestingSaga : Saga<RequestResponseRequestingSaga.RequestResponseRequestingSagaData>,
+                IAmStartedByMessages<InitiateRequestingSaga>,
+                IHandleMessages<ResponseFromOtherSaga>
+            {
+                public Context TestContext { get; set; }
+
+                public Task Handle(InitiateRequestingSaga message, IMessageHandlerContext context)
+                {
+                    return context.SendLocal(new RequestToRespondingSaga
+                    {
+                        SomeIdThatTheResponseSagaCanCorrelateBackToUs = Data.CorrIdForResponse //wont be needed in the future
+                    });
+                }
+
+                public Task Handle(ResponseFromOtherSaga message, IMessageHandlerContext context)
+                {
+                    TestContext.DidRequestingSagaGetTheResponse = true;
+
+                    MarkAsComplete();
+
+                    return Task.FromResult(0);
+                }
+
+                protected override void ConfigureHowToFindSaga(SagaPropertyMapper<RequestResponseRequestingSagaData> mapper)
+                {
+                    mapper.ConfigureMapping<InitiateRequestingSaga>(m => m.Id).ToSaga(s => s.CorrIdForResponse);
+                    mapper.ConfigureMapping<ResponseFromOtherSaga>(m => m.SomeCorrelationId).ToSaga(s => s.CorrIdForResponse);
+                }
+
+                public class RequestResponseRequestingSagaData : ContainSagaData
+                {
+                    public virtual Guid CorrIdForResponse { get; set; } //wont be needed in the future
+                }
+            }
+
+            public class RequestResponseRespondingSaga : Saga<RequestResponseRespondingSaga.RequestResponseRespondingSagaData>,
+                IAmStartedByMessages<RequestToRespondingSaga>,
+                IHandleMessages<SendReplyFromNonInitiatingHandler>
+            {
+                public Context TestContext { get; set; }
+
+                public Task Handle(RequestToRespondingSaga message, IMessageHandlerContext context)
+                {
+                    return context.SendLocal(new SendReplyFromNonInitiatingHandler
+                    {
+                        SagaIdSoWeCanCorrelate = Data.CorrIdForRequest
+                    });
+                }
+
+                public Task Handle(SendReplyFromNonInitiatingHandler message, IMessageHandlerContext context)
+                {
+                    //reply to originator must be used here since the sender of the incoming message the timeoutmanager and not the requesting saga
+                    return ReplyToOriginator(context, new ResponseFromOtherSaga //change this line to Bus.Reply(new ResponseFromOtherSaga  and see it fail
+                    {
+                        SomeCorrelationId = Data.CorrIdForRequest //wont be needed in the future
+                    });
+                }
+
+                protected override void ConfigureHowToFindSaga(SagaPropertyMapper<RequestResponseRespondingSagaData> mapper)
+                {
+                    mapper.ConfigureMapping<RequestToRespondingSaga>(m => m.SomeIdThatTheResponseSagaCanCorrelateBackToUs).ToSaga(s => s.CorrIdForRequest);
+                    //this line is just needed so we can test the non initiating handler case
+                    mapper.ConfigureMapping<SendReplyFromNonInitiatingHandler>(m => m.SagaIdSoWeCanCorrelate).ToSaga(s => s.CorrIdForRequest);
+                }
+
+                public class RequestResponseRespondingSagaData : ContainSagaData
+                {
+                    public virtual Guid CorrIdForRequest { get; set; }
+                }
+            }
+        }
+
+        public class InitiateRequestingSaga : ICommand
+        {
+            public InitiateRequestingSaga()
+            {
+                Id = Guid.NewGuid();
+            }
+
+            public Guid Id { get; set; }
+        }
+
+        public class RequestToRespondingSaga : ICommand
+        {
+            public Guid SomeIdThatTheResponseSagaCanCorrelateBackToUs { get; set; }
+        }
+
+        public class ResponseFromOtherSaga : IMessage
+        {
+            public Guid SomeCorrelationId { get; set; }
+        }
+
+        public class SendReplyFromNonInitiatingHandler : ICommand
+        {
+            public Guid SagaIdSoWeCanCorrelate { get; set; }
         }
     }
 }
